+++ ld_script.txt
@@ -316,7 +316,6 @@ SECTIONS {
         asm/code_809C5C4.o(.text);
         src/code_809D148.o(.text);
         src/ground_script.o(.text);
-        src/code_80A14E8.o(.text);
         src/ground_script_2.o(.text);
         src/code_80A26CC.o(.text);
         asm/code_80A26CC_1.o(.text);
